+++ src/tasks/PSFactory.Pester.Tasks/tasks/Invoke-Pester.pester.build.ps1
@@ -183,7 +183,7 @@ task Invoke_Pester_Tests_v4 {
         ForEach-Object -Process {
             if ($_ -is [System.Collections.Hashtable])
             {
-                Convert-FactoryHashtableToString -Hashtable $_
+                Format-FactoryHashtable -Hashtable $_
             }
             else
             {
@@ -199,7 +199,7 @@ task Invoke_Pester_Tests_v4 {
     "`tExclude Cov.  = $($ExcludeFromCodeCoverage -join ', ')"
     "`tModuleVersion = $ModuleVersion"
 
-    $osShortName = Get-FactoryOperatingSystemShortName
+    $osShortName = Get-FactoryOS
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
 
@@ -319,7 +319,7 @@ task Invoke_Pester_Tests_v4 {
             {
                 foreach ($scriptItem in $PesterScript)
                 {
-                    Write-Build -Color 'DarkGray' -Text "      ... $(Convert-FactoryHashtableToString -Hashtable $scriptItem)"
+                    Write-Build -Color 'DarkGray' -Text "      ... $(Format-FactoryHashtable -Hashtable $scriptItem)"
 
                     $pesterParams.Script += $scriptItem
                 }
@@ -369,7 +369,7 @@ task Fail_Build_If_Pester_Tests_Failed {
 
     "`tPester Output Folder     = '$PesterOutputFolder'"
 
-    $osShortName = Get-FactoryOperatingSystemShortName
+    $osShortName = Get-FactoryOS
 
     $GetCodeCoverageThresholdParameters = @{
         RuntimeCodeCoverageThreshold = $CodeCoverageThreshold
@@ -454,7 +454,7 @@ task Invoke_Pester_Tests_v5 {
         $null = New-Item -Path $PesterOutputFolder -ItemType 'Directory' -Force -ErrorAction 'Stop'
     }
 
-    $osShortName = Get-FactoryOperatingSystemShortName
+    $osShortName = Get-FactoryOS
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
 
@@ -946,7 +946,7 @@ task Pester_If_Code_Coverage_Under_Threshold {
         $PesterOutputFolder = Join-Path -Path $OutputDirectory -ChildPath $PesterOutputFolder
     }
 
-    $osShortName = Get-FactoryOperatingSystemShortName
+    $osShortName = Get-FactoryOS
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
 
@@ -1037,7 +1037,7 @@ task Upload_Test_Results_To_AppVeyor -If { (property BuildSystem 'unknown') -eq 
         $null = New-Item -Path $PesterOutputFolder -ItemType Directory -Force -ErrorAction 'Stop'
     }
 
-    $osShortName = Get-FactoryOperatingSystemShortName
+    $osShortName = Get-FactoryOS
 
     $powerShellVersion = 'PSv.{0}' -f $PSVersionTable.PSVersion
 
@@ -1094,7 +1094,7 @@ task Pester_Run_Times {
     $getPesterOutputFileFileNameParameters = @{
         ProjectName       = $ProjectName
         ModuleVersion     = $ModuleVersion
-        OsShortName       = Get-FactoryOperatingSystemShortName
+        OsShortName       = Get-FactoryOS
         PowerShellVersion = ('PSv.{0}' -f $PSVersionTable.PSVersion)
     }