+++ CHANGELOG.md
@@ -687,7 +687,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Made Convert-FactoryHashtableToString public function.
+- Made Format-FactoryHashtable public function.
 - Refactored a lot of Path resolution into PSFactory public function for consitency and re-usability.
 - Updated the Tasks to use those PSFactory functions.
 - Updated Get-BuiltModuleVersion to support $BuiltModuleSubdirectory as per #239.